Reboot project
This is an attempt at restarting the project using a different approach:
using a custom GitHub Actions workflow to build & publish the site.

As far as I can tell, this is a recent option that's still in Beta
(https://github.blog/changelog/2022-07-27-github-pages-custom-github-actions-workflows-beta/),
but it definitely looks to be a better alternative to the previous
iteration of this project.

There's some examples and documentation on how to do this:
- https://docs.github.com/en/pages/getting-started-with-github-pages/configuring-a-publishing-source-for-your-github-pages-site
- https://github.com/actions/configure-pages
- https://github.com/actions/upload-pages-artifact
- https://github.com/actions/deploy-pages